Correct the alias vertex data types.
Vertex locations need to be unsigned byte rather than byte (GL is funy
with that). s and t need to be at least short, and since the normal index
is embedded in the st vector, it needs to be the same type. With this, my
test tetrahedrons seem to be working.

+++ include/QF/GLSL/qf_alias.h
@@ -33,7 +33,7 @@
 
 typedef struct aliasvrt_s {
 	byte        vertex[3];
-	byte        stn[3];
+	short       stn[3];
 } aliasvrt_t;
 
 void R_InitAlias (void);
